third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.proxy-same.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify same-origin Partitioned Popins proxies only have access to postMessage and closed methods.
  assert_equals: expected "Closed,Then," but got "Closed,Blur,OnBlur,Opener,Length,Name,AnonymousIndex,AnonymousName,CustomMethod,CustomAttributeGet,CustomAttributeSet,Then,"
Harness: the test ran to completion.
